QR-Encoder-Decoder/Rndastech/decoder/requirements.txt
Pillow==10.0.0
pyzbar==0.1.9
